New functions + bug in setCentroid
New functions: expanded the regionprops_withKnownCentroids to trigger a
custom function over the PixelIntensity values

Bug: Identified a bug in setting centroids in the GUI. Under certain
circumstances, the GUI allows the user to place a centroid in the edge
of the image (value 0). I placed a line in the setCentroid to set the
minimum centroid to 0. This doesn't yet solve the case for the other
edge of the image. I also modified the getCentroids and getValues
functions in the CentroidTimeseries object to look both in col and row
when validating centroids.