Bugs in lineage tree
Two bugs:
1. Arising from rounding errors in identifying sister cells based on
centroid information.
2. A gui error, that doesn't move the division centroid in
centroidsDivisions when editing a track. This is due to the fact that
when annotating a division or death event the 'value' parameter was
assumed to have been modified, but this was not the case. I hope to be
able to consolidate all annotations in a future release to stop using
individual objects for each annotation type.